+++ ext_localconf.php
@@ -1,6 +1,8 @@
 <?php
 defined('TYPO3_MODE') || die();
 
+$GLOBALS['TYPO3_CONF_VARS']['FE']['addRootLineFields'] .= ',tx_schema_webpagetype';
+
 $GLOBALS['TYPO3_CONF_VARS']['SC_OPTIONS']['t3lib/class.t3lib_pagerenderer.php']['render-postProcess'][] =
     \Brotkrueml\Schema\Hook\PageRenderer\PostProcessHook::class . '->execute';